feature dist tests
